requirements.txt
ccxt>=2.0.0
pandas>=1.3.0
numpy>=1.20.0
python-dotenv>=0.19.0
ta>=0.9.0
tradingview-ta>=3.3.0
matplotlib>=3.5.0
backtesting>=0.3.3
pytest>=7.0.0
requests>=2.27.0
schedule>=1.1.0
pymongo>=4.0.0
pytest-mock>=3.7.0
python-dateutil>=2.8.2
pytz>=2022.1
email-validator>=1.2.0 